Check if current id is same as old id

+++ src/arcgis/geoprocessing/_job.py
@@ -972,6 +972,16 @@ class OMJob(GPJob):
                                 pass
                         mission_json["items"].update({key: {}})
 
+            itemid = None
+            if self._item:                
+                item_props = json.loads(self._item)
+                if "serviceProperties" in item_props.keys():
+                    if "itemProperties" in item_props.keys():
+                        if "itemId" in item_props["itemProperties"].keys():
+                            itemid = item_props["itemProperties"]["itemId"]
+                elif "itemId" in item_props.keys():
+                    itemid = item_props["itemId"]
+
             if item_name == "dsm" or item_name == "dtm" or item_name == "ortho":
                 if "items" in mission_json.keys():
                     for key in mission_json["items"].keys():
@@ -981,14 +991,15 @@ class OMJob(GPJob):
                                 isinstance(item_info, dict)
                                 and "itemId" in item_info.keys()
                             ):
-                                item_object = mission._gis.content.get(
-                                    item_info["itemId"]
-                                )
-                                try:
-                                    if item_object:
-                                        deleted = item_object.delete()
-                                except:
-                                    pass
+                                if item_info["itemId"] != itemid:
+                                    item_object = mission._gis.content.get(
+                                        item_info["itemId"]
+                                    )
+                                    try:
+                                        if item_object:
+                                            deleted = item_object.delete()
+                                    except:
+                                        pass
                             mission_json["items"].update({key: {}})
                             if key in mission_json["jobs"].keys():
                                 mission_json["jobs"].update({key: {"checked": False}})
